record exception

+++ src/main/java/com/example/demo/controller/PublishController.java
@@ -24,7 +24,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -44,7 +44,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam(value = "id", required = false, defaultValue = "0") Integer id,
+            @RequestParam(value = "id", required = false, defaultValue = "0") Long id,
             HttpServletRequest request,
             Model model){
